refactor(app): add `infrastructure` directory, move `core`
