+++ src/app/infrastructure/core/pipes/pipe.module.ts
