+++ src/app/infrastructure/core/pipes/safe-html.pipe.ts
